Moved BMI to its own object and created an object for profiles
The program used to run only on the view controller, I moved it to its own class. Profile passes in the BMI object now and the Profile object is called by the view controller. The app currently still only does the BMI calculation but profiles will been a near addition.